requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-ads==23.1.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2
